requirements.txt
git+https://github.com/huggingface/diffusers.git
git+https://github.com/huggingface/accelerate.git
git+https://github.com/huggingface/peft.git
transformers>=4.50.0
huggingface_hub
sentencepiece
torchvision
numpy
Pillow
